+++ src/config/recommended.ts
@@ -5,6 +5,7 @@ module.exports = {
 		'close-attr': 'error',
 		'close-order': 'error',
 		'deprecated': 'error',
+		'element-name': 'error',
 		'element-permitted-content': 'error',
 		'no-dup-id': 'error',
 		'no-trailing-whitespace': 'error',